update 'game' page to 'play game' and add game landing as 'game' page instead
