--- hooks/recipeSwitch.js
@@ -1,14 +1,0 @@
-import { useState } from "react";
-
-export default function useRecipeSwitch() {
-    const [showIngredientList, setShowIngredientList] = useState(true)
-    const [showDirections, setShowDirections] = useState(false)
-
-    const op = () => {
-        setShowIngredientList(!showIngredientList);
-        setShowDirections(!showDirections);
-    }
-
-
-    return {op, showDirections, showIngredientList}
-}
