+++ hooks/showGame.js
@@ -1,0 +1,9 @@
+import { useState } from "react";
+
+const displayGame = () => {
+    const [showGame, setShowGame] = useState(false);
+
+    return{showGame, setShowGame};
+}
+
+export default displayGame;
