+++ styles/PlayGame.module.css
@@ -1,0 +1,398 @@
+.playGame__page {
+    position: relative;
+    display: flex;
+    justify-content: center;
+}
+
+/* Game Over & Cleared */
+.gameOver, .cleared  {
+    display: flex;
+    justify-content: center;
+    position: absolute;
+    z-index: 1;
+    background-color: #034806bf;
+    width: 100vw;
+    height: 100vh;
+    align-items: center;
+}
+
+.gameOverDisplay, .clearedDisplay {
+    width: 100%;
+    max-width: 350px;
+    height: 600px;
+    position: relative;
+    background-color: var(--color-white);
+    border-radius: 20px;
+}
+
+.gameOverDesc, .clearedDesc {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    gap: 20px;
+}
+
+.gameOverDesc > h1 {
+    padding-top: 60px;
+    font-family: var(--secondary-font);
+    color: var(--color-red);
+}
+
+.gameOverDesc > p, .clearedDesc > p {
+    width: 180px;
+    text-align: center;
+}
+
+.gameOverDesc > h3 {
+    color: var(--color-avocado);
+}
+
+.clearedDesc > h1 {
+    padding-top: 60px;
+    font-family: var(--secondary-font);
+    color: var(--color-avocado);
+}
+
+.clearedDesc > h3 {
+    color: var(--color-red);
+}
+
+.goOptions {
+    display: flex;
+    gap: 60px;
+}
+
+.playerSelection {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    gap: 7px;
+    font-size: 20px;
+    cursor: pointer;
+}
+
+/* HeadLine */
+.navBar {
+    max-width: 900px;
+    width: 100%;
+    display: flex;
+    justify-content: center;
+    padding: 30px 30px 0 30px;
+}
+
+.navBarOption {
+    max-width: 700px;
+    width: 100%;
+    display: flex;
+    justify-content: space-between;
+    color: var(--color-black);
+}
+
+.seeTutorial {
+    display: flex;
+    align-items: center;
+}
+
+.gameOptiones {
+    position: relative;
+    display: flex;
+    flex-direction: column;
+    gap: 10px;
+    align-items: end;
+}
+
+.gameLife {
+    position: absolute;
+    top: 50px;
+    right: 20px;
+    color: var(--color-pure-white);
+}
+
+.headlines {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    gap: 10px;
+}
+
+.subHealine {
+    font-size: var(--body-font-size-desktop);
+    font-family: var(--primary-font);
+}
+
+.mainHeadline {
+    font-size: var(--main-headline-font-size-desktop);
+    font-family: var(--secondary-font);
+    letter-spacing: 1px;
+    color: var(--color-red);
+}
+
+/* Main Game Lauout */
+.game_container {
+    width: 100%;
+}
+
+.gameLayout {
+    width: 100%;
+    height: 100vh;
+    position: relative;
+    display: flex;
+    flex-direction: column;
+    justify-content: space-between;
+    align-items: center;
+    padding-bottom: 120px;
+}
+
+/* Stage 1 */
+.selectIngredients {
+    width: 100%;
+    display: flex;
+    justify-content: center;
+}
+
+.ingredientsTypes {
+    max-width: 300px;
+    width: 100%;
+    display: flex;
+    justify-content: space-between;
+}
+
+.ingredients {
+    display: flex;
+    flex-direction: column;
+    gap: 15px;
+}
+
+.containerPickIngredients {
+    width: 100%;
+    display: flex;
+    justify-content: center;
+}
+
+.ingredientsList {
+    max-width: 300px;
+    width: 100%;
+    display: flex;
+    flex-wrap: wrap;
+    gap: 25px;
+}
+
+.ingredientsSections {
+    display: flex;
+    flex-direction: column;
+    gap: 15px;
+}
+
+.ingredientImages {
+    border: 2px solid var(--color-lightgray);
+    border-radius: 5px;
+    cursor: pointer;
+    transition: transform .2s;
+}
+
+.ingredientImages:hover {
+    transform: scale(1.1);
+}
+
+.next__button {
+    display: flex;
+    justify-content: center;
+}
+
+/* Stage 2 */
+.gameLayoutSecondStage {
+    padding-top: 80px;
+}
+.stageTwoImages {
+    position: absolute;
+    z-index: -1;
+    border: 2px solid var(--color-lightgray);
+    border-radius: 5px;
+}
+
+.stageTwoImagesRice {
+    top: 210px;
+    left: 32%;
+}
+
+.stageTwoImagesRiceAnimated {
+    top: 635px;
+    left: 50%;
+    opacity: 0;
+    animation-name: addRice;
+    animation-duration: 1.5s;
+}
+@keyframes addRice {
+    0%{
+        top: 210px;
+        left: 32%;
+        opacity: 1;
+    }
+    50% {
+        opacity: 0.1;
+        width: 50px;
+        height: 50px;
+    } 
+    100% {
+        top: 700px;
+        left: 50%;
+        width: 0px;
+        height: 0px;
+        opacity: 0;
+    }
+}
+
+.stageTwoImagesBeef {
+    top: 210px;
+    right: 32%;
+}
+
+.stageTwoImagesBeefAnimated {
+    top: 700px;
+    right: 50%;
+    opacity: 0;
+    animation-name: addBeef;
+    animation-duration: 1.5s;
+}
+@keyframes addBeef {
+    0%{
+        top: 210px;
+        right: 32%;
+        opacity: 1;
+    }
+    50% {
+        opacity: 0.1;
+        width: 50px;
+        height: 50px;
+    } 
+    100% {
+        top: 700px;
+        right: 50%;
+        width: 0px;
+        height: 0px;
+        opacity: 0;
+    }
+}
+
+.stageTwoImagesVeggies {
+    top: 385px;
+    left: 32%;
+}
+
+.stageTwoImagesVeggiesAnimated {
+    top: 700px;
+    left: 50%;
+    opacity: 0;
+    animation-name: addVeggies;
+    animation-duration: 1.5s;
+}
+@keyframes addVeggies {
+    0%{
+        top: 385px;
+        left: 32%;
+        opacity: 1;
+    }
+    50% {
+        opacity: 0.1;
+        width: 50px;
+        height: 50px;
+    } 
+    100% {
+        top: 700px;
+        left: 50%;
+        width: 0px;
+        height: 0px;
+        opacity: 0;
+    }
+}
+
+.stageTwoImagesEgg {
+    top: 385px;
+    right: 32%;
+}
+
+.stageTwoImagesEggAnimated {
+    top: 700px;
+    right: 50%;
+    opacity: 0;
+    animation-name: addEgg;
+    animation-duration: 1.5s;
+}
+@keyframes addEgg {
+    0%{
+        top: 385px;
+        right: 32%;
+        opacity: 1;
+    }
+    50% {
+        opacity: 0.1;
+        width: 50px;
+        height: 50px;
+    } 
+    100% {
+        top: 700px;
+        right: 50%;
+        width: 0px;
+        height: 0px;
+        opacity: 0;
+    }
+}
+
+.letsCookButton {
+    padding-top: 340px;
+    animation-name: fadeIn;
+    animation-duration: 1.5s;
+}
+@keyframes fadeIn {
+    0%{
+        opacity: 0;
+    } 
+    100% {
+        opacity: 1;
+    }
+}
+
+/* Stage 3 */
+.gameLayoutThirdStage {
+    padding-top: 80px;
+    padding-bottom: 130px;
+}
+
+.thirdStageOptions {
+    display: flex;
+    gap: 50px;
+} 
+
+.thirdStageOptionDisplay {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    gap: 10px;
+}
+
+/* Mobile Size */
+@media (max-width: 500px) {
+    .navBar {
+        max-width: 415px;
+    }
+    .stageTwoImages {
+        width: 120px;
+        height: 120px;
+    }
+    .stageTwoImagesRice {
+        top: 200px;
+        left: 18%;
+    } 
+    .stageTwoImagesVeggies {
+        top: 340px;
+        left: 18%;
+    }
+    .stageTwoImagesBeef {
+        top: 200px;
+        right: 18%;
+    } 
+    .stageTwoImagesEgg {
+        top: 340px;
+        right: 18%;
+    }
+}
